MS_2-T_2 - Add refactoring changes to reg page
