+++ frontend/tochkarik/src/pages/RegisterPage/RegisterPage.css
@@ -1,0 +1,18 @@
+.register-container {
+    display: flex;
+    flex-direction: column;
+
+    align-items: center;
+}
+
+.register-container form {
+    display: flex;
+    flex-direction: column;
+
+    justify-content: space-around;
+
+    width: 300px;
+    height: 200px;
+
+    margin: auto;
+}
