+++ frontend/tochkarik/src/routes/UnauthenticatedRoutes.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Routes, Route, Navigate } from 'react-router-dom';
 
 import LoginPage from "../pages/LoginPage.jsx";
-import RegistrationPage from "../pages/RegistrationPage.jsx";
+import RegistrationPage from "../pages/RegisterPage/RegistrationPage.jsx";
 
 function UnauthenticatedRoutes({ setAuthToken }) {
     return (
